Add an error code for read buffer overflow
This adds an error code to deal with the case where the HTTP request
does not fit in the read buffer.  Previously the code returned an
`UnexpectedEof` error, which seems somewhat confusing, as this implies
there was a problem with the client's request.

The code now returns an HTTP 431 error if a client sends a request that
does not fit in the read buffer.