Fix #96. Add method to shutdown embassy listeners
Instead of blocking listening on sockets there is a
timeout. Each loop run it checks if the worker
still needs to be running and exits the loop if not.

This is not backwards compatible with previous versions
as it introduces a new paramter.
To be decided if we want a new method, and keep old
behaviour if not passed.

Did not implement for tokio, as the socket loop seems
to be implemented in app.